Restructured folders and components to be more structured by feauture, layout and ui components. Removed userpreview css file and added that class name to index.scss. Changed height for bottomnavigation. Also renamed some components to be more clear. Refactored header title handling
