+++ frontend/sass/_index.scss
@@ -15,3 +15,8 @@
   padding-top: 90px;
   padding-bottom: 90px;
 }
+
+.profile-img {
+  width: 50px;
+  border-radius: 50%;
+}
