+++ frontend/sass/_sticky-footer.scss
@@ -1,5 +1,6 @@
 .bottom-nav {
   position: fixed;
+  height: 80px;
   bottom: 0;
   width: 100%;
   background-color: rgb(255, 255, 255);
@@ -17,12 +18,14 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  font-size: 20px;
+  font-size: 22px;
   transition: all 0.25s ease;
 }
 
 .nav-item.active {
-  font-size: 25px;
+  font-size: 28px;
+  width: 50px;
+  height: 50px;
   background: #41445c;
   color: white;
   transform: translateY(-8px);
