+++ frontend/src/App.tsx
@@ -1,11 +1,10 @@
 import { Outlet, useLocation } from "react-router-dom";
 import { Toaster } from "react-hot-toast";
-import NavigationBar from "./components/Navigation/NavigationBar";
-import Header from "./components/Header/Header";
+import NavigationBar from "./components/layout/BottomNavigation";
+import Header from "./components/layout/Header";
 import { Container } from "react-bootstrap";
 
 function App() {
-  // scroll to top when the route changes
   useLocation();
   window.scrollTo({ top: 0, left: 0, behavior: "instant" });
   return (
