--- frontend/src/components/UserPreview.css
@@ -1,9 +1,0 @@
-.profile-img {
-  width: 50px;
-  border-radius: 50%;
-}
-
-p {
-  color:rgba(0, 0, 0, 0.342);
-  margin-bottom: 0 !important;
-}
